Fix reset to default columns
